#3 anti_fraud исправил замечания сергея

+++ antifraud/src/main/java/com/bank/antifraud/dto/SuspiciousCardTransferDto.java
@@ -1,7 +1,13 @@
 package com.bank.antifraud.dto;
 
 import jakarta.persistence.Column;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -10,9 +16,9 @@ import lombok.*;
 @Setter
 @ToString
 public class SuspiciousCardTransferDto {
-    int id;
+    long id;
 
-    int cardTransferId;
+    long cardTransferId;
 
     boolean blocked;
